netsec-dashboard/requirements.txt
fastapi
uvicorn[standard]
snowflake-connector-python
pydantic
python-dotenv
httpx
wireshark